Fix performance regression using Cython 3.0
In Cython 3, `cdef` functions defaults to raising exceptions
as opposed to in Cython 0.2x. Here we add `noexcept` to all `nogil`
`cdef` functions.

Even with `noexcept` typed memoryviews will raise an exception which
requires GIL and reduces performance. Deep in the documentation I
found the compiler directive `initializedcheck` which if set to False
disables the exception.