Defer repository refresh to first use.
Before this commit, the default repository was refreshed upon importing
the datasets module. If the user has no internet connectivity, the
import would block for 30 seconds while trying to resolve the DNS before
failing.

This commit defer the refresh until the first access of a repository,
giving an overall better experience offline.

Unfortunately, even with this fix we still block for 30 seconds before
loading the cached repository.

I'm still considering ways of improving this.