Fix a major bug in ShapeletTreeClassifier/Regressor
Most likely a fix for #88.

When fitting a shapelet tree, we evaluate the goodness of every
shapelet. However, by mistake the criterion was not reset between
different shapelets. As a result, we always selected the first
shapelet irrespective of number of shapelets we inspected. As
a result, the trees where built as if n_shapelets where set to 1.

This fix correctly resets the criterion.