Add new parameter `impurity_equality_tolerance`.
The `impurity_equality_tolerance` parameter determines the tolerance for
considering two impurity values as equal. When the absolute difference
between the current impurity and the best impurity encountered so far is
less than or equal to this tolerance, the tree will also consider the
size of the distance gap between the splits.

By default using the separation gap is disabled.